back-end/requirements.txt
Flask==2.1.0
pandas==2.2.3
numpy==2.0.2
openai==0.27.0
folium==0.12.1
python-dotenv==0.19.2
flask-caching==1.10.1
Werkzeug==2.0.3
scikit-learn
gunicorn==20.1.0
flask-restx
flask_sqlalchemy
flask_migrate
SQLAlchemy==1.4.49
flask_cors

